cheri: drivers: net: Fix compile in the rest of drivers/net/
- Increase alignment for structs that contain pointers
- Add uintptr_t casts.

+++ drivers/net/usb/lan78xx.c
@@ -453,7 +453,7 @@ struct lan78xx_net {
 	unsigned int		maxpacket;
 	struct timer_list	stat_monitor;
 
-	unsigned long		data[5];
+	uintptr_t		data[5];
 
 	int			link_on;
 	u8			mdix_ctrl;
